experiments/requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Embeddings and ML
google-generativeai>=0.8.0
openai>=1.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0

# NLP and Entity Extraction
spacy>=3.7.0

# Graph processing
networkx>=3.1
hypernetx>=2.0.0

# Vector storage
faiss-cpu>=1.7.4
chromadb>=0.4.0

# Memory system libraries (baselines)
letta>=1.0.0
mem0ai>=1.0.0
langchain>=0.2.0
langchain-community>=0.2.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0

# Evaluation and analysis
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
httpx>=0.25.0
tenacity>=8.2.0
aiohttp>=3.9.0

# Type checking
mypy>=1.5.0
types-requests>=2.31.0
